# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/password-reset/request": {
      "post": {
        "description": "Public. Answers 202 with the same acknowledgement whether or not the address has an account — telling the two apart would turn this route into an account-existence oracle. Delivery happens out of band.",
        "operationId": "CredentialRecoveryController_requestPasswordReset_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequestDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Request accepted. Same answer for an unknown address.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcknowledgementDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body, or an email that is not an address.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Request a password reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password-reset/confirm": {
      "post": {
        "description": "Public — the reset token is the credential. Sets the new password and revokes the token. Single-use: replaying it answers 401. Existing sessions are not silently kept alive, so the caller logs in again afterwards.",
        "operationId": "CredentialRecoveryController_confirmPasswordReset_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcknowledgementDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body, or a password that fails the policy.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Unknown, expired or already-used reset token.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Confirm a password reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/magic-link/request": {
      "post": {
        "description": "Public. Answers 202 with the same acknowledgement whether or not the address has an account — telling the two apart would turn this route into an account-existence oracle. Delivery happens out of band.",
        "operationId": "CredentialRecoveryController_requestMagicLink_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequestDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Request accepted. Same answer for an unknown address.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcknowledgementDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body, or an email that is not an address.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Request a magic-link login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/magic-link/consume": {
      "post": {
        "description": "Public — the magic-link token is the credential, so no password is involved. Opens a session: the access token comes back in the body, the refresh token only as an `httpOnly` path-scoped cookie. Single-use: replaying answers 401.",
        "operationId": "CredentialRecoveryController_consumeMagicLink_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session opened. Refresh token set as a cookie, never in the body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Unknown, expired or already-used magic-link token.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Log in with a magic link",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tu Seguridad API",
    "description": "Space-scoped video surveillance API. Every route outside `auth` and `health` resolves resources inside the caller space, taken from the bearer token, so an id from another space answers 404 rather than a result. Writes that change what the system watches are admin-only. Failures answer `{ statusCode, code, message }` — branch on `code`, not on the message.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404,
            "description": "HTTP status, the same value as the response status line."
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "INVALID_ZONE",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "UPSTREAM_ERROR",
              "UPSTREAM_TIMEOUT",
              "INTERNAL_ERROR"
            ],
            "example": "NOT_FOUND",
            "description": "Stable machine-readable code. Branch on this, not on the message."
          },
          "message": {
            "type": "string",
            "example": "Camera 7f3a not found",
            "description": "Human-readable reason. Never carries credentials, tokens or snapshot URLs."
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
...
      "AccessTokenDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Bearer token for the `Authorization` header. Short-lived — rotate through `POST /auth/refresh` when it expires.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI..."
          }
        },
        "required": [
          "accessToken"
        ]
      },
...
      "EmailRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "owner@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "AcknowledgementDto": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "accepted"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque token from the delivered link"
          },
          "password": {
            "type": "string",
            "minLength": 12
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "CredentialTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque token from the delivered link"
          }
        },
        "required": [
          "token"
        ]
      },
...
    }
  }
}